+++ crates/unbase/src/network/transport/simulator.rs
@@ -29,21 +29,20 @@ struct SimEvent {
 impl SimEvent {
     pub fn deliver (self) {
         //println!("# SimEvent.deliver" );
-        if let Some(to_slab) = self.dest.upgrade() {
 
-            /* let memo = &self.memoref.get_memo_if_resident().unwrap();
-            println!("Simulator.deliver FROM {} TO {} -> {}({:?}): {:?} {:?} {:?}",
-                &self.from_slabref.slab_id,
-                &to_slab.id,
-                &self.memoref.id,
-                &self.memoref.subject_id,
-                &memo.body,
-                &memo.parents.memo_ids(),
-                &self.memoref.peerlist.read().unwrap().slab_ids()
-            );*/
-            let owned_slabref = &self.from_slabref.clone_for_slab(&to_slab);
-            self.memoref.clone_for_slab( &owned_slabref, &to_slab, true );
-        }
+        /* let memo = &self.memoref.get_memo_if_resident().unwrap();
+        println!("Simulator.deliver FROM {} TO {} -> {}({:?}): {:?} {:?} {:?}",
+            &self.from_slabref.slab_id,
+            &to_slab.id,
+            &self.memoref.id,
+            &self.memoref.subject_id,
+            &memo.body,
+            &memo.parents.memo_ids(),
+            &self.memoref.peerlist.read().unwrap().slab_ids()
+        );*/
+        let slabref = self.dest.agent.localize_slabref(&self.from_slabref);
+        self.dest.agent.localize_memoref( &self.memoref, &slabref, true );
+
         // we all have to learn to deal with loss sometime
     }
 }
