+++ crates/unbase/src/slab/memo/mod.rs
@@ -134,56 +134,7 @@ impl Memo {
         }
         return false;
     }
-    pub fn clone_for_slab (&self, from_slabref: &SlabRef, to_slab: &SlabHandle, peerlist: &MemoPeerList) -> Memo {
-        assert!(from_slabref.owning_slab_id == to_slab.id, "Memo clone_for_slab owning slab should be identical");
-
-        //println!("Slab({}).Memo.clone_for_slab(memo: {}, from: {}, to: {}, peers: {:?})", self.owning_slab_id, self.id, from_slabref.slab_id, to_slab.id, peerlist );
-        to_slab.reconstitute_memo(
-            self.id,
-            self.subject_id,
-            self.parents.clone_for_slab(from_slabref, to_slab, false),
-            self.body.clone_for_slab(from_slabref, to_slab),
-            from_slabref,
-            peerlist
-        ).0
-    }
 }
 
 impl MemoBody {
-    fn clone_for_slab(&self, from_slabref: &SlabRef, to_slab: &SlabHandle ) -> MemoBody {
-        assert!(from_slabref.owning_slab_id == to_slab.id, "MemoBody clone_for_slab owning slab should be identical");
-
-        match self {
-            &MemoBody::SlabPresence{ ref p, ref r } => {
-                MemoBody::SlabPresence{
-                    p: p.clone(),
-                    r: match r {
-                        &Some(ref root_mrh) => {
-                            Some(root_mrh.clone_for_slab(from_slabref, to_slab, true))
-                        }
-                        &None => None
-                    }
-                }
-            },
-            &MemoBody::Relation(ref rssh) => {
-                MemoBody::Relation(rssh.clone_for_slab(from_slabref, to_slab))
-            }
-            &MemoBody::Edit(ref hm) => {
-                MemoBody::Edit(hm.clone())
-            }
-            &MemoBody::FullyMaterialized{ ref v, ref r } => {
-                MemoBody::FullyMaterialized{ v: v.clone(), r: r.clone_for_slab(from_slabref, to_slab)}
-            }
-            &MemoBody::PartiallyMaterialized{ ref v, ref r } => {
-                MemoBody::PartiallyMaterialized{ v: v.clone(), r: r.clone_for_slab(from_slabref, to_slab)}
-            }
-            &MemoBody::Peering(memo_id, subject_id, ref peerlist) => {
-                MemoBody::Peering(memo_id,subject_id,peerlist.clone_for_slab(to_slab))
-            }
-            &MemoBody::MemoRequest(ref memo_ids, ref slabref) =>{
-                MemoBody::MemoRequest(memo_ids.clone(), slabref.clone_for_slab(to_slab))
-            }
-        }
-
-    }
 }
