--- crates/unbase/src/slab/mod.rs
@@ -1,378 +1,0 @@
-use std::collections::hash_map::Entry;
-
-pub use self::common_structs::*;
-pub use self::slabref::{SlabRef,SlabRefInner};
-pub use self::memoref::{MemoRef,MemoRefInner,MemoRefPtr};
-pub use self::memo::{MemoId,Memo,MemoInner,MemoBody};
-pub use self::memoref::serde as memoref_serde;
-pub use self::memo::serde as memo_serde;
-
-
-use crate::subject::SubjectId;
-use crate::memorefhead::*;
-use crate::context::{Context,WeakContext};
-use crate::network::{Network,Transmitter,TransportAddress};
-
-use std::sync::{Arc,RwLock,Mutex};
-use std::sync::mpsc;
-use std::sync::mpsc::channel;
-use std::fmt;
-
-mod state;
-mod agent;
-mod common_structs;
-mod handle;
-
-mod memo;
-mod slabref;
-mod memoref;
-
-pub use handle::SlabHandle;
-
-pub type SlabId = u32;
-
-use crate::slab::agent::SlabAgent;
-use futures::{StreamExt, Future};
-use futures::future::RemoteHandle;
-
-type DispatcherFuture = impl Future<Output = u32> + Send;
-type Dispatcher = impl Fn() -> DispatcherFuture;
-
-#[derive(Clone)]
-pub struct Slab{
-    pub id: SlabId,
-    agent: Arc<SlabAgent>,
-    net: Network,
-    pub my_ref: SlabRef,
-    dispatch_channel: futures::channel::mpsc::Sender<MemoRef>,
-    dispatcher: Arc<RemoteHandle<Dispatcher>>,
-}
-
-impl Slab {
-    pub fn new(net: &Network) -> Slab {
-        let id = net.generate_slab_id();
-
-        let my_ref_inner = SlabRefInner {
-            slab_id: id,
-            owning_slab_id: id, // I own my own ref to me, obviously
-            presence: RwLock::new(vec![]), // this bit is just for show
-            tx: Mutex::new(Transmitter::new_blackhole(id)),
-            return_address: RwLock::new(TransportAddress::Local),
-        };
-
-        let my_ref = SlabRef(Arc::new(my_ref_inner));
-        // TODO: figure out how to reconcile this with the simulator
-
-        let (dispatch_tx_channel, dispatch_rx_channel) = futures::channel::mpsc::channel::<MemoRef>(10);
-
-        let agent = Arc::new(SlabAgent::new( net, my_ref.clone() ));
-
-        let agent2 = agent.clone();
-        let dispatcher  = crate::util::task::spawn_with_handle( async move || {
-            let mut dispatch_rx_channel = dispatch_rx_channel;
-            while let Some(memoref) = dispatch_rx_channel.next().await {
-                agent2.recv_memoref(memoref);
-            }
-        });
-
-        let me = Slab{
-            id,
-            dispatch_channel: dispatch_tx_channel,
-            dispatcher: Arc::new(dispatcher),
-            net: net.clone(),
-            my_ref: my_ref,
-            agent
-        };
-
-        net.register_local_slab(me.handle() );
-
-        net.conditionally_generate_root_index_seed(&me);
-
-        me
-    }
-    pub fn handle(&self) -> SlabHandle {
-        SlabHandle::new(self.my_ref.clone())
-    }
-    pub fn get_root_index_seed (&self) -> Option<MemoRefHead> {
-        self.net.get_root_index_seed(&self.handle())
-    }
-    pub fn create_context (&self) -> Context {
-        Context::new(self)
-    }
-    pub fn subscribe_subject (&self, subject_id: u64, context: &Context) {
-        //println!("Slab({}).subscribe_subject({})", self.id, subject_id );
-        let weakcontext : WeakContext = context.weak();
-
-        match self.subject_subscriptions.write().unwrap().entry(subject_id){
-            Entry::Occupied(mut e) => {
-                e.get_mut().push(weakcontext)
-            }
-            Entry::Vacant(e) => {
-                e.insert(vec![weakcontext]);
-            }
-        }
-        return;
-    }
-    pub fn unsubscribe_subject (&self,  subject_id: u64, context: &Context ){
-        if let Some(subs) = self.subject_subscriptions.write().unwrap().get_mut(&subject_id) {
-            let weak_context = context.weak();
-            subs.retain(|c| {
-                c.cmp(&weak_context)
-            });
-            return;
-        }
-    }
-    pub fn memo_wait_channel (&self, memo_id: MemoId ) -> mpsc::Receiver<Memo> {
-        let (tx, rx) = channel::<Memo>();
-
-        match self.memo_wait_channels.lock().unwrap().entry(memo_id) {
-            Entry::Vacant(o)       => { o.insert( vec![tx] ); }
-            Entry::Occupied(mut o) => { o.get_mut().push(tx); }
-        };
-
-        rx
-    }
-    pub fn generate_subject_id(&self) -> SubjectId {
-        let mut counters = self.counters.write().unwrap();
-        counters.last_subject_id += 1;
-        (self.id as u64).rotate_left(32) | counters.last_subject_id as u64
-    }
-    fn _memo_durability_score( &self, _memo: &Memo ) -> u8 {
-        // TODO: devise durability_score algo
-        //       Should this number be inflated for memos we don't care about?
-        //       Or should that be a separate signal?
-
-        // Proposed factors:
-        // Estimated number of copies in the network (my count = count of first order peers + their counts weighted by: uptime?)
-        // Present diasporosity ( my diasporosity score = mean peer diasporosity scores weighted by what? )
-        0
-    }
-    pub fn check_memo_waiters ( &self, memo: &Memo) {
-        match self.memo_wait_channels.lock().unwrap().entry(memo.id) {
-            Entry::Occupied(o) => {
-                for channel in o.get() {
-                    // we don't care if it worked or not.
-                    // if the channel is closed, we're scrubbing it anyway
-                    channel.send(memo.clone()).ok();
-                }
-                o.remove();
-            },
-            Entry::Vacant(_) => {}
-        };
-    }
-    pub fn do_peering(&self, memoref: &MemoRef, origin_slabref: &SlabRef) {
-
-        let do_send = if let Some(memo) = memoref.get_memo_if_resident(){
-            // Peering memos don't get peering memos, but Edit memos do
-            // Abstracting this, because there might be more types that don't do peering
-            memo.does_peering()
-        }else{
-            // we're always willing to do peering for non-resident memos
-            true
-        };
-
-        if do_send {
-            // That we received the memo means that the sender didn't think we had it
-            // Whether or not we had it already, lets tell them we have it now.
-            // It's useful for them to know we have it, and it'll help them STFU
-
-            // TODO: determine if peering memo should:
-            //    A. use parents at all
-            //    B. and if so, what should be should we be using them for?
-            //    C. Should we be sing that to determine the peered memo instead of the payload?
-            //println!("MEOW {}, {:?}", my_ref );
-
-            let peering_memoref = self.new_memo(
-                None,
-                memoref.to_head(),
-                MemoBody::Peering(
-                    memoref.id,
-                    memoref.subject_id,
-                    memoref.get_peerlist_for_peer(&self.my_ref, Some(origin_slabref.slab_id))
-                )
-            );
-            origin_slabref.send( &self.my_ref, &peering_memoref );
-        }
-
-    }
-    pub fn handle_memo_from_other_slab( &self, memo: &Memo, memoref: &MemoRef, origin_slabref: &SlabRef ){
-        //println!("Slab({}).handle_memo_from_other_slab({})", self.id, memo.id );
-
-        match memo.body {
-            // This Memo is a peering status update for another memo
-            MemoBody::SlabPresence{ p: ref presence, r: ref opt_root_index_seed } => {
-
-                match opt_root_index_seed {
-                    &Some(ref root_index_seed) => {
-
-                        // HACK - this should be done inside the deserialize
-                        for memoref in root_index_seed.iter() {
-                            memoref.update_peer(origin_slabref, MemoPeeringStatus::Resident);
-                        }
-
-                        self.net.apply_root_index_seed( &presence, root_index_seed, &self.my_ref );
-                    }
-                    &None => {}
-                }
-
-                let mut reply = false;
-                if let &None = opt_root_index_seed {
-                    reply = true;
-                }
-
-                if reply {
-                    if let Ok(mentioned_slabref) = self.slabref_from_presence( presence ) {
-                        // TODO: should we be telling the origin slabref, or the presence slabref that we're here?
-                        //       these will usually be the same, but not always
-
-                        let my_presence_memoref = self.new_memo_basic(
-                            None,
-                            memoref.to_head(),
-                            MemoBody::SlabPresence{
-                                p: self.presence_for_origin( origin_slabref ),
-                                r: self.get_root_index_seed()
-                            }
-                        );
-
-                        origin_slabref.send( &self.my_ref, &my_presence_memoref );
-
-                        let _ = mentioned_slabref;
-                        // needs PartialEq
-                        //if mentioned_slabref != origin_slabref {
-                        //   mentioned_slabref.send( &self.my_ref, &my_presence_memoref );
-                        //}
-                    }
-                }
-            }
-            MemoBody::Peering(memo_id, subject_id, ref peerlist ) => {
-                let (peered_memoref,_had_memo) = self.assert_memoref( memo_id, subject_id, peerlist.clone(), None );
-
-                // Don't peer with yourself
-                for peer in peerlist.iter().filter(|p| p.slabref.0.slab_id != self.id ) {
-                    peered_memoref.update_peer( &peer.slabref, peer.status.clone());
-                }
-            },
-            MemoBody::MemoRequest(ref desired_memo_ids, ref requesting_slabref ) => {
-
-                if requesting_slabref.0.slab_id != self.id {
-                    for desired_memo_id in desired_memo_ids {
-                        if let Some(desired_memoref) = self.memorefs_by_id.read().unwrap().get(&desired_memo_id) {
-
-                            if desired_memoref.is_resident() {
-                                requesting_slabref.send(&self.my_ref, desired_memoref)
-                            } else {
-                                // Somebody asked me for a memo I don't have
-                                // It would be neighborly to tell them I don't have it
-                                self.do_peering(&memoref,requesting_slabref);
-                            }
-                        }else{
-                            let peering_memoref = self.new_memo(
-                                None,
-                                MemoRefHead::from_memoref(memoref.clone()),
-                                MemoBody::Peering(
-                                    *desired_memo_id,
-                                    None,
-                                    MemoPeerList::new(vec![MemoPeer{
-                                        slabref: self.my_ref.clone(),
-                                        status: MemoPeeringStatus::NonParticipating
-                                    }])
-                                )
-                            );
-                            requesting_slabref.send(&self.my_ref, &peering_memoref)
-                        }
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    // Counters,stats, reporting
-    pub fn count_of_memorefs_resident( &self ) -> u32 {
-        self.memorefs_by_id.read().unwrap().len() as u32
-    }
-    pub fn count_of_memos_received( &self ) -> u64 {
-        self.counters.read().unwrap().memos_received as u64
-    }
-    pub fn count_of_memos_reduntantly_received( &self ) -> u64 {
-        self.counters.read().unwrap().memos_redundantly_received as u64
-    }
-    pub fn peer_slab_count (&self) -> usize {
-        self.peer_refs.read().unwrap().len() as usize
-    }
-
-    pub fn new_memo_basic (&self, subject_id: Option<SubjectId>, parents: MemoRefHead, body: MemoBody) -> MemoRef {
-        self.new_memo(subject_id, parents, body)
-    }
-    pub fn new_memo_basic_noparent (&self, subject_id: Option<SubjectId>, body: MemoBody) -> MemoRef {
-        self.new_memo(subject_id, MemoRefHead::new(), body)
-    }
-    pub fn remotize_memo_ids( &self, memo_ids: &[MemoId] ) -> Result<(),String>{
-        //println!("# Slab({}).remotize_memo_ids({:?})", self.id, memo_ids);
-
-        let mut memorefs : Vec<MemoRef> = Vec::with_capacity(memo_ids.len());
-
-        {
-            let memorefs_by_id = self.memorefs_by_id.read().unwrap();
-            for memo_id in memo_ids.iter() {
-                if let Some(memoref) = memorefs_by_id.get(memo_id) {
-                    memorefs.push( memoref.clone() )
-                }
-            }
-        }
-
-        for memoref in memorefs {
-            self.remotize_memoref(&memoref)?;
-        }
-
-        Ok(())
-    }
-    // should this be a function of the slabref rather than the owning slab?
-    pub fn presence_for_origin (&self, origin_slabref: &SlabRef ) -> SlabPresence {
-        // Get the address that the remote slab would recogize
-        SlabPresence {
-            slab_id: self.id,
-            address: origin_slabref.get_return_address(),
-            lifetime: SlabAnticipatedLifetime::Unknown
-        }
-    }
-    pub fn slabref_from_local_slab(&self, peer_slab: &SlabHandle) -> SlabRef {
-
-        //let args = TransmitterArgs::Local(&peer_slab);
-        let presence = SlabPresence{
-            slab_id: peer_slab.id,
-            address: TransportAddress::Local,
-            lifetime: SlabAnticipatedLifetime::Unknown
-        };
-
-        self.assert_slabref(peer_slab.id, &vec![presence])
-    }
-    pub fn slabref_from_presence(&self, presence: &SlabPresence) -> Result<SlabRef,&str> {
-        match presence.address {
-            TransportAddress::Simulator  => {
-                return Err("Invalid - Cannot create simulator slabref from presence")
-            }
-            TransportAddress::Local      => {
-                return Err("Invalid - Cannot create local slabref from presence")
-            }
-            _ => {
-                unimplemented!()
-            }
-        };
-
-
-        //let args = TransmitterArgs::Remote( &presence.slab_id, &presence.address );
-
-        Ok(self.assert_slabref( presence.slab_id, &vec![presence.clone()] ))
-    }
-}
-
-impl fmt::Debug for Slab {
-    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-        fmt.debug_struct("Slab")
-            .field("slab_id", &self.id)
-            .field("peer_refs", &self.peer_refs)
-            .field("memo_refs", &self.memorefs_by_id)
-            .finish()
-    }
-}
